chore: address comments
Context: https://github.com/unoplatform/uno/pull/22935

Note: `WebView.md` links to `xref:Uno.Features.NativeAOT`, which is
from #22935.
